+++ client/src/components/modal.jsx
@@ -1,0 +1,27 @@
+const modalRoot = document.getElementById('modal-root');
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+class Modal extends React.Component {
+    constructor(props) {
+        super(props);
+        this.el = document.createElement('div');
+    }
+
+    componentDidMount() {
+        modalRoot.appendChild(this.el);
+    }
+
+    componentWillUnmount() {
+        modalRoot.removeChild(this.el);
+    }
+
+    render() {
+        return ReactDOM.createPortal(
+            this.props.children,
+            this.el,
+        );
+    }
+}
+
+export default Modal;
