+++ package.json
@@ -32,6 +32,7 @@
     "babel-preset-react": "^6.24.1",
     "bluebird": "^3.7.2",
     "body-parser": "^1.19.0",
+    "dateformat": "^3.0.3",
     "express": "^4.17.1",
     "faker": "^4.1.0",
     "jquery": "^3.4.1",
